requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.3
brotli==1.1.0
pillow==11.0.0
